+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Date: XXX
 * ReactionPlaneFlow: Returns also the angle, not only the flow value
 * Jetscape: Possibility to read in parton output files with optional `particletype` parameter in constructor
 * Particle: Include changes for parton read in from Jetscape class, multiply quark charges by 3 to make them integers
+* MultiParticlePtCorrelations: Class to compute higher order mean transverse momentum correlations and the corresponding cumulants
 
 ### Fixed
 * QCumulantFlow and LeeYangZeroFlow: Fix bug in the calculation of the differential flow